dp-client/requirements.txt
metadata-client
psycopg2-binary>=2.9,<3.0

